@@ -3,7 +3,7 @@ project(incppect)
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS "on")
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
 
